+++ etherlink/kernel_latest/revm/src/storage/block.rs
@@ -37,7 +37,7 @@ mod test {
     use crate::Error;
     use revm::primitives::B256;
     use tezos_evm_runtime::runtime::MockKernelHost;
-    use tezos_smart_rollup_host::runtime::Runtime;
+    use tezos_smart_rollup_host::storage::StorageV1;
 
     #[test]
     fn test_malformed_hash() {