+++ src/kernel_tx_demo/kernel/src/transactions/utils.rs
@@ -33,7 +33,7 @@ pub(crate) fn read_large_store_chunk<'a>(
 #[cfg(test)]
 mod tests {
     use tezos_smart_rollup_host::path::RefPath;
-    use tezos_smart_rollup_host::runtime::Runtime;
+    use tezos_smart_rollup_host::storage::StorageV1;
     use tezos_smart_rollup_mock::MockHost;
 
     use crate::{transactions::utils::read_large_store_chunk, MAX_ENVELOPE_CONTENT_SIZE};