+++ src/kernel_sdk/CHANGES.md
@@ -42,6 +42,8 @@
 - Add `tezos-smart-rollup-keyspace` crate which defines the `KeySpace` high-level durable storage API
 - Split `write_debug` from `Runtime` to new `HostDebug` supertrait.  Will allow kernels to restrict scope to just to
   host's debug messaging capabilities if desired.
+- Split `store_` functions from `Runtime` to new `StorageV1` supertrait. Will allow kernels to restrict scope to just
+  WASM-pvm flavour durable storage capabilities if desired.
 
 ### Installer client/kernel